requirements.txt
fastapi
uvicorn
pandas
python-dotenv
langchain_groq
langdetect
fuzzywuzzy
deep-translator
httpx
googletrans
asyncio
mapbox
google.generativeai
faiss-cpu